osc sync range fix, serial send and receive amp comp calibratin offset

+++ params.h
@@ -37,7 +37,7 @@ bool RESONANCEAmpCompensation = true;
 bool VCFADSRRestart = true;
 bool VCAADSRRestart = true;
 int16_t VCFKeytrack = 0;
-int8_t oscSyncMode = 0;
+uint16_t oscSyncMode = 0;
 
 int8_t ADSR3ToOscSelect;
